COL-1781, refactor to consolidate utils; progress on EditActiveFabricObject
