+++ squiggy/models/whiteboard_element.py
@@ -81,8 +81,8 @@ class WhiteboardElement(Base):
         return asset_whiteboard_element
 
     @classmethod
-    def update(cls, element, whiteboard_element_id, asset_id=None):
-        whiteboard_element = cls.query.filter_by(id=whiteboard_element_id).first()
+    def update(cls, element, uuid, whiteboard_id, asset_id=None):
+        whiteboard_element = cls.query.filter_by(uuid=uuid, whiteboard_id=whiteboard_id).first()
         whiteboard_element.asset_id = asset_id
         whiteboard_element.element = element
         db.session.add(whiteboard_element)
