+++ src/store/whiteboarding/constants.ts
